Add: update list with months
