+++ src/routes/budget/+page.svelte
@@ -6,6 +6,7 @@
 	import BudgetItemsList from '$lib/components/AllBudgetList.svelte';
 	import CategoryBudgetList from '$lib/components/CategoryBudgetList.svelte';
 	import { setBudget } from '$lib/components/functions';
+	import MonthSlider from '$lib/components/MonthSlider.svelte';
 
 	interface Props {
 		data: BudgetData
@@ -20,16 +21,17 @@
 </script>
 
 <main class="flex h-full flex-col items-center justify-start gap-2 mb-16">
+	<MonthSlider budgetItems={data.budgetItems} />
 	<div class="flex h-full w-full flex-col items-center justify-between gap-6 p-4">
 		{#if !data.budgetItems.length}
 			<span class="p-8 text-center text-lg"
 				>This months budget is empty, add an item to start budgeting</span
 			>
 		{:else}
-			<DonutChart budgetItems={data.budgetItems}/>
+			<DonutChart/>
 			<div class="flex w-full flex-col items-center justify-center gap-2 px-5">
 				{#if !clickedChartInfo.budgetCategory}
-					<BudgetItemsList budgetItems={data.budgetItems} />
+					<BudgetItemsList />
 				{:else if clickedChartInfo.budgetCategory && clickedChartInfo.budgetSum}
 					<CategoryBudgetList />
 				{/if}
